Moving generic_work service tests to model spec
