--- spec/hydra/works/services/generic_work/get_related_objects_spec.rb
@@ -1,45 +1,0 @@
-require 'spec_helper'
-
-describe Hydra::Works::GetRelatedObjectsFromGenericWork do
-
-  subject { Hydra::Works::GenericWork::Base.new }
-
-  let(:object1) { Hydra::PCDM::Object.new }
-  let(:object2) { Hydra::PCDM::Object.new }
-
-  let(:generic_work1) { Hydra::Works::GenericWork::Base.new }
-  let(:generic_work2) { Hydra::Works::GenericWork::Base.new }
-  let(:generic_file1) { Hydra::Works::GenericFile::Base.new }
-
-  describe '#call' do
-    context 'with generic files and works' do
-      before do
-        Hydra::Works::AddGenericWorkToGenericWork.call( subject, generic_work1 )
-        Hydra::Works::AddGenericWorkToGenericWork.call( subject, generic_work1 )
-        Hydra::Works::AddGenericFileToGenericWork.call( subject, generic_file1 )
-      end
-
-      it 'should return empty array when only generic files and generic works are aggregated' do
-        expect(Hydra::Works::GetRelatedObjectsFromGenericWork.call( subject )).to eq []
-      end
-
-      it 'should only return related objects' do
-        Hydra::Works::AddRelatedObjectToGenericWork.call( subject, object2 )
-        expect(Hydra::Works::GetRelatedObjectsFromGenericWork.call( subject )).to eq [object2]
-      end
-
-      it 'should return related objects of various types' do
-        Hydra::Works::AddRelatedObjectToGenericWork.call( subject, generic_work2 )
-        Hydra::Works::AddRelatedObjectToGenericWork.call( subject, generic_file1 )
-        Hydra::Works::AddRelatedObjectToGenericWork.call( subject, object1 )
-        related_objects = Hydra::Works::GetRelatedObjectsFromGenericWork.call( subject )
-        expect( related_objects.include? object1 ).to be true
-        expect( related_objects.include? generic_work2 ).to be true
-        expect( related_objects.include? generic_file1 ).to be true
-        expect( related_objects.size ).to eq 3
-      end
-   end
-  end
-end
-
-
